pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.64.0",
    "chromadb>=1.0.20",
    "openai>=1.100.0",
    "pymupdf>=1.26.3",
    "pypdf2>=3.0.1",
    "python-docx>=1.2.0",
    "pyyaml>=6.0.2",
    "requests>=2.32.4",
    "rich>=14.1.0",
    "stringzilla>=3.12.6",
    "typer>=0.16.1",
]
